// apps/desktop/src/ipc/generated/SaveState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AppError } from "./AppError";
import type { SaveStatus } from "./SaveStatus";

export type SaveState = { status: SaveStatus, lastSavedAt: number | null, error: AppError | null, pendingOperations: number, };